Backend/config.py
import os
from google.cloud import secretmanager
# print("Config file loaded")
# print(f"PINECONE_API_KEY: {os.environ.get('PINECONE_API_KEY')}")
print("Config file loaded securely using Google Secret Manager")

def get_secret(name):
    client = secretmanager.SecretManagerServiceClient()
    project_id = os.getenv("GCP_PROJECT")  # Set this during deployment
    secret_name = f"projects/{project_id}/secrets/{name}/versions/latest"
    response = client.access_secret_version(request={"name": secret_name})
    return response.payload.data.decode("UTF-8")


PINECONE_API_KEY = os.environ.get("PINECONE_API_KEY")
OPENAI_API_KEY = os.environ.get("OPENAI_API_KEY")
COHERE_API_KEY = os.environ.get("COHERE_API_KEY")

# Vector Database Settings
VECTOR_DB_NAME = "pinecone"  
PINECONE_INDEX_NAME = "data28k"

# Embedding Model Settings
EMBEDDING_MODEL = "Snowflake/snowflake-arctic-embed-m-v1.5"

# LLM Settings
LLM_PROVIDER = "openai" 
LLM_MODEL = "gpt-4o-mini"  #  "gpt-3.5-turbo"
LLM_TEMPERATURE = 0.2
LLM_MAX_TOKENS = 500

# Retrieval Settings
TOP_K_RETRIEVE = 10

# Reranker Settings
USE_RERANKER = True  # Set to False to disable reranking
RERANKER_PROVIDER = "cohere"  
RERANKER_MODEL = "rerank-english-v2.0"
TOP_K_RERANK = 5

# NEU Assistant Prompt Template
NEU_PROMPT_TEMPLATE = """
CONTEXT:
{context}

QUESTION:
{question}

You are a knowledgeable assistant specializing in Northeastern University (NEU) information. Your purpose is to provide friendly, straightforward responses that sound natural and conversational.

Guidelines:
- Present information about NEU as factual knowledge without mentioning "context," "provided information," or any references to your information sources in the main body of your answer
- Extract source URLs found at the beginning of each information chunk
- After your complete answer, include a "Sources" section that lists all relevant URLs
- If you lack sufficient information to answer fully, simply state: "I don't have enough information about this topic. Please contact Northeastern University directly for more details."  
- Never fabricate information
- Maintain professional language regardless of user input
- If faced with inappropriate language, respond professionally while addressing the underlying question if legitimate

Every information chunk begins with a URL. Include these URLs only in your "Sources" section at the end of your response, never in the main answer.
This is important, so if you answer from any document make sure you put sources at the end of answer, never miss it. 

Your responses should sound natural and helpful, as if you're simply sharing knowledge about Northeastern University without revealing how you obtained that information.
"""
