Backend/requirements.txt
fastapi
uvicorn
openai
pinecone-client
cohere
torch
transformers
langchain
langchain-community
langchain-core
google-cloud-secret-manager
